+++ src/games/flappyBird/flappyBird.cpp
@@ -1,7 +1,9 @@
 #include "flappyBird.hpp"
+
+#include "zcollide.hpp"
 #include "zmath.hpp"
 
-FlappyBird::FlappyBird(ZWindow& zwindow, ma_engine& audio, Collision& collision, Push& vertex) : zwindow(zwindow), audio(audio), collision(collision), vertex(vertex) {}
+FlappyBird::FlappyBird(ZWindow& zwindow, ma_engine& audio, Push& vertex) : zwindow(zwindow), audio(audio), vertex(vertex) {}
 
 float flappyBirdSpeedY = 0.f;
 bool flappyBirdDead = false;
@@ -20,7 +22,7 @@ void FlappyBird::tick() {
                 sprites[0].textureIndex = 0;
 
                 playSound(&audio, "assets/sounds/chirp.mp3");
-                createSprite(squareModel, 3, 0.f, 0.f, .5f, .2f, 0.f, 1.f, 1.f, 1.f, 1.f);
+                createSprite(squareModel, 3, 0.f, 0.f, .5f, .2f, 0.f);
                 //sprites[sprites.size() - 1].setText("ZDEV", 0, 32.f);
                 vertex.cameraZoom[0] -= .025f;
                 vertex.cameraZoom[1] -= .025f;
@@ -52,7 +54,7 @@ void FlappyBird::tick() {
         sprites[index + 1].position[0] = sprites[index].position[0];
         sprites[index + 1].position[1] = sprites[index].position[1] - 2.f;
 
-        if (flappyBirdStarted && !flappyBirdDead && (collision.checkSquareCollision(sprites[index], sprites[0]) || collision.checkSquareCollision(sprites[index + 1], sprites[0]))) {
+        if (flappyBirdStarted && !flappyBirdDead && (zcollide_checkSquareCollision(sprites[index], sprites[0]) || zcollide_checkSquareCollision(sprites[index + 1], sprites[0]))) {
             flappyBirdDead = true;
             flappyBirdSpeedY = -1.5f;
             sprites[0].rotation = 60.f;