+++ src/games/slimeAttack/slimeAttack.hpp
@@ -1,19 +1,17 @@
 #pragma once
 
 #include "zwindow.hpp"
-#include "collision.hpp"
 #include "zengine.hpp"
 
 #include "slimeAttackEnemies.hpp"
 
 struct SlimeAttack {
 public:
-    SlimeAttack(ZWindow& zwindow, Collision& collision);
+    SlimeAttack(ZWindow& zwindow);
 
     void tick();
     void knockback(float origin);
 private:
     SlimeAttackEnemies slimeAttackEnemies;
-    Collision& collision;
     ZWindow& zwindow;
 };